+++ packages/veui/src/components/Transfer/_SelectedPanel.vue
@@ -28,7 +28,7 @@
     <veui-tree
       v-if="showMode === 'tree'"
       :datasource="items"
-      :expands.sync="localExpands"
+      :expanded.sync="localExpanded"
       class="veui-transfer-selected-tree"
       @click="remove"
     >
@@ -149,27 +149,27 @@ export default {
     filter: Function,
     placeholder: String,
     isSelectable: Boolean,
-    expands: Array,
+    expanded: Array,
     icons: Object
   },
   data () {
     return {
       flattenOptions: [],
-      localExpands: []
+      localExpanded: []
     }
   },
   watch: {
-    expands: {
+    expanded: {
       handler (val, oldVal) {
         if (!isEqual(val, oldVal)) {
-          this.localExpands = clone(this.expands)
+          this.localExpanded = clone(this.expanded)
         }
       },
       immediate: true
     },
-    localExpands (val, oldVal) {
+    localExpanded (val, oldVal) {
       if (!isEqual(val, oldVal)) {
-        this.$emit('update:expands', val)
+        this.$emit('update:expanded', val)
       }
     },
     datasource: {
@@ -223,14 +223,14 @@ export default {
       )
     },
     toggle (option) {
-      let expands = clone(this.localExpands)
-      let index = expands.indexOf(option.value)
+      let expanded = clone(this.localExpanded)
+      let index = expanded.indexOf(option.value)
       if (index > -1) {
-        expands.splice(index, 1)
+        expanded.splice(index, 1)
       } else {
-        expands.push(option.value)
+        expanded.push(option.value)
       }
-      this.localExpands = expands
+      this.localExpanded = expanded
     },
     removeAll () {
       this.$emit('removeall')